backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
openpyxl==3.1.5
pandas==2.2.2
whitenoise==6.7.0
dj-database-url==2.2.0
